refactor out StripeConnectWebViewContainerImpl
