+++ connect/api/connect.api
@@ -14,14 +14,6 @@ public final class com/stripe/android/connect/AccountOnboardingProps$Creator : a
 	public synthetic fun newArray (I)[Ljava/lang/Object;
 }
 
-public final class com/stripe/android/connect/AccountOnboardingView : com/stripe/android/connect/StripeComponentView, com/stripe/android/connect/webview/StripeConnectWebViewContainer {
-	public static final field $stable I
-	public synthetic fun getListener ()Lcom/stripe/android/connect/StripeEmbeddedComponentListener;
-	public fun initialize (Lcom/stripe/android/connect/EmbeddedComponentManager;Lcom/stripe/android/connect/AccountOnboardingListener;Lcom/stripe/android/connect/AccountOnboardingProps;)V
-	public synthetic fun initialize (Lcom/stripe/android/connect/EmbeddedComponentManager;Lcom/stripe/android/connect/StripeEmbeddedComponentListener;Lcom/stripe/android/connect/ComponentProps;)V
-	public synthetic fun setListener (Lcom/stripe/android/connect/StripeEmbeddedComponentListener;)V
-}
-
 public final class com/stripe/android/connect/BuildConfig {
 	public static final field BUILD_TYPE Ljava/lang/String;
 	public static final field DEBUG Z
@@ -43,14 +35,6 @@ public final class com/stripe/android/connect/EmptyProps$Creator : android/os/Pa
 	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
 	public final fun newArray (I)[Lcom/stripe/android/connect/EmptyProps;
 	public synthetic fun newArray (I)[Ljava/lang/Object;
-}
-
-public final class com/stripe/android/connect/PayoutsView : com/stripe/android/connect/StripeComponentView, com/stripe/android/connect/webview/StripeConnectWebViewContainer {
-	public static final field $stable I
-	public synthetic fun getListener ()Lcom/stripe/android/connect/StripeEmbeddedComponentListener;
-	public fun initialize (Lcom/stripe/android/connect/EmbeddedComponentManager;Lcom/stripe/android/connect/PayoutsListener;Lcom/stripe/android/connect/EmptyProps;)V
-	public synthetic fun initialize (Lcom/stripe/android/connect/EmbeddedComponentManager;Lcom/stripe/android/connect/StripeEmbeddedComponentListener;Lcom/stripe/android/connect/ComponentProps;)V
-	public synthetic fun setListener (Lcom/stripe/android/connect/StripeEmbeddedComponentListener;)V
 }
 
 public abstract interface class com/stripe/android/connect/StripeComponentController$OnDismissListener {
